Refs #3712. Removed unused profiles.
