+++ client.config
@@ -1,7 +1,3 @@
-PROFILE_CREATE_ENTITIES_REF=TRUE
-PROFILE_CREATE_ENTITIES_XML=TRUE
-PROFILE_READ_ACCESS=TRUE
-PROFILE_WRITE_ACCESS=TRUE
 PROFILE_DISCOVERY=TRUE
 PROFILE_UDP_TRANSPORT=TRUE
 PROFILE_TCP_TRANSPORT=TRUE
