+++ examples/ShapesDemo/CMakeLists.txt
@@ -19,14 +19,10 @@ endif()
 
 project(ShapeDemoClient)
 
-if(NOT PROFILE_CREATE_ENTITIES_REF OR
-   NOT PROFILE_CREATE_ENTITIES_XML OR
-   NOT PROFILE_READ_ACCESS OR
-   NOT PROFILE_WRITE_ACCESS OR
-   NOT PROFILE_UDP_TRANSPORT OR
+if(NOT PROFILE_UDP_TRANSPORT OR
    NOT PROFILE_TCP_TRANSPORT OR
    NOT PROFILE_SERIAL_TRANSPORT)
-    message(WARNING "Can not compile test: The profiles PROFILE_CREATE_ENTITIES_REF, PROFILE_CREATE_ENTITIES_XML, PROFILE_READ_ACCESS, PROFILE_WRITE_ACCESS, PROFILE_UDP_TRANSPORT, PROFILE_TCP_PROFILE and PROFILE_SERIAL_TRANSPORT must be enabled.")
+    message(WARNING "Can not compile test: The PROFILE_UDP_TRANSPORT, PROFILE_TCP_PROFILE and PROFILE_SERIAL_TRANSPORT must be enabled.")
 else()
     add_executable(${PROJECT_NAME} main.c ShapeType.c)
     if(MSVC OR MSVC_IDE)
