+++ examples/SubscribeHelloWorld/CMakeLists.txt
@@ -19,10 +19,8 @@ endif()
 
 project(SubscribeHelloWorldClient)
 
-if(NOT PROFILE_CREATE_ENTITIES_XML OR
-   NOT PROFILE_READ_ACCESS OR
-   NOT PROFILE_UDP_TRANSPORT)
-    message(WARNING "Can not compile example: The profiles, PROFILE_CREATE_ENTITIES_XML, PROFILE_READ_ACCESS, and PROFILE_UDP_TRANSPORT must be enabled.")
+if(NOT PROFILE_UDP_TRANSPORT)
+    message(WARNING "Can not compile example: The PROFILE_UDP_TRANSPORT must be enabled.")
 else()
     add_executable(${PROJECT_NAME} main.c HelloWorld.c)
     if(MSVC OR MSVC_IDE)
