+++ include/uxr/client/config.h.in
@@ -20,11 +20,8 @@
 #define UXR_CLIENT_VERSION_MICRO @PROJECT_VERSION_PATCH@
 #define UXR_CLIENT_VERSION_STR "@PROJECT_VERSION@"
 
-#cmakedefine PROFILE_CREATE_ENTITIES_REF
-#cmakedefine PROFILE_CREATE_ENTITIES_XML
-#cmakedefine PROFILE_READ_ACCESS
-#cmakedefine PROFILE_WRITE_ACCESS
 #cmakedefine PROFILE_DISCOVERY
+
 #cmakedefine PROFILE_UDP_TRANSPORT
 #cmakedefine PROFILE_TCP_TRANSPORT
 #cmakedefine PROFILE_SERIAL_TRANSPORT
