+++ include/uxr/client/core/session/common_create_entities.h
@@ -12,24 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _SRC_C_PROFILE_SESSION_COMMON_CREATE_ENTITIES_INTERNAL_H_
-#define _SRC_C_PROFILE_SESSION_COMMON_CREATE_ENTITIES_INTERNAL_H_
+#ifndef _UXR_CLIENT_CORE_SESSION_COMMON_CREATE_ENTITIES_H_
+#define _UXR_CLIENT_CORE_SESSION_COMMON_CREATE_ENTITIES_H_
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
-#include <uxr/client/profile/session/common_create_entities.h>
+#include <uxr/client/core/session/session.h>
 
-struct CREATE_Payload;
-
-uint16_t uxr_common_create_entity(uxrSession* session, uxrStreamId stream_id,
-                              uxrObjectId object_id, uint16_t xml_ref_size, uint8_t mode,
-                              struct CREATE_Payload* payload);
+UXRDLLAPI uint16_t uxr_buffer_delete_entity(uxrSession* session, uxrStreamId stream_id, uxrObjectId object_id);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif //_SRC_C_PROFILE_SESSION_COMMON_CREATE_ENTITIES_INTERNAL_H_
+#endif //_UXR_CLIENT_CORE_SESSION_COMMON_CREATE_ENTITIES_H_
