+++ include/uxr/client/core/session/create_entities_ref.h
@@ -13,15 +13,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _UXR_CLIENT_PROFILE_SESSION_CREATE_ENTITIES_REF_H_
-#define _UXR_CLIENT_PROFILE_SESSION_CREATE_ENTITIES_REF_H_
+#ifndef _UXR_CLIENT_CORE_SESSION_CREATE_ENTITIES_REF_H_
+#define _UXR_CLIENT_CORE_SESSION_CREATE_ENTITIES_REF_H_
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
-#include <uxr/client/profile/session/common_create_entities.h>
+#include <uxr/client/core/session/common_create_entities.h>
 
 UXRDLLAPI uint16_t uxr_buffer_create_participant_ref(uxrSession* session, uxrStreamId stream_id,
                                                   uxrObjectId object_id, int16_t domain_id,
@@ -43,4 +43,4 @@ UXRDLLAPI uint16_t uxr_buffer_create_datareader_ref(uxrSession* session, uxrStre
 }
 #endif
 
-#endif //_UXR_CLIENT_PROFILE_SESSION_CREATE_ENTITIES_REF_H_
+#endif //_UXR_CLIENT_CORE_SESSION_CREATE_ENTITIES_REF_H_
