+++ src/c/core/session/common_create_entities.c
@@ -1,8 +1,8 @@
 #include "common_create_entities_internal.h"
-#include "../../core/session/session_internal.h"
-#include "../../core/session/session_info_internal.h"
-#include "../../core/session/submessage_internal.h"
-#include "../../core/serialization/xrce_protocol_internal.h"
+#include "session_internal.h"
+#include "session_info_internal.h"
+#include "submessage_internal.h"
+#include "../serialization/xrce_protocol_internal.h"
 
 //==================================================================
 //                              PUBLIC
