+++ test/memory/memory_map/complete_profile.config
@@ -1,6 +1,3 @@
-PROFILE_CREATE_ENTITIES_REF=TRUE
-PROFILE_CREATE_ENTITIES_XML=TRUE
-PROFILE_READ_ACCESS=TRUE
 PROFILE_UDP_TRANSPORT=TRUE
 PROFILE_TCP_TRANSPORT=TRUE
 PROFILE_SERIAL_TRANSPORT=TRUE
