+++ test/memory/memory_map/tcp_profile.config
@@ -1,6 +1,3 @@
-PROFILE_CREATE_ENTITIES_REF=FALSE
-PROFILE_CREATE_ENTITIES_XML=FALSE
-PROFILE_READ_ACCESS=FALSE
 PROFILE_UDP_TRANSPORT=FALSE
 PROFILE_TCP_TRANSPORT=TRUE
 PROFILE_SERIAL_TRANSPORT=FALSE
