+++ test/memory/memory_map/xml_profile.config
@@ -1,6 +1,3 @@
-PROFILE_CREATE_ENTITIES_REF=FALSE
-PROFILE_CREATE_ENTITIES_XML=TRUE
-PROFILE_READ_ACCESS=FALSE
 PROFILE_UDP_TRANSPORT=FALSE
 PROFILE_TCP_TRANSPORT=FALSE
 PROFILE_SERIAL_TRANSPORT=FALSE
